Moved DocumentStatusFactories into DocumentStatusFactory folder
